Moved structures for main platform into header file.
